Move netty operations to transport package
